requirements.txt
aiogram==2.19
aiohttp==3.8.3
aioschedule==0.5.2
aiosignal==1.3.1
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
blockchain==1.4.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.1
coinbase==2.1.0
cryptocompare==0.7.0
cryptography==3.3.2
decorator==4.4.2
emoji==0.5.4
enum-compat==0.0.3
Flask==1.1.2
flask-crontab==0.1.2
frozenlist==1.3.3
future==0.18.2
idna==2.9
importlib-metadata==6.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==6.0.4
oauthlib==3.2.2
paho-mqtt==1.6.1
psycopg2-binary==2.8.5
pyaml==21.10.1
pycparser==2.20
pycryptodome==3.9.7
pyMeterBus==0.8.3
pypng==0.20220715.0
PyQRCode==1.2.1
pyserial==3.5
pyTelegramBotAPI==3.6.7
python-dotenv==0.13.0
python-telegram-bot==12.6.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rfeed==1.1.1
simplejson==3.18.4
six==1.14.0
SQLAlchemy==1.3.16
tornado==6.0.4
tqdm==4.66.1
tweepy==4.12.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==1.0.1
yarl==1.8.2
zipp==3.13.0
